Fix handling include paths when they don't include SCSS
The filtering I wrote to sanitize the include paths was a bit dumb: if
the include paths were purely non-SCSS related files (like, say, a list
of Ruby files), then it would filter down to an empty list, which would
be get passed to SCSS itself and be interpreted as "lint everything".

The fix here is to behave more similarly to other engines: `["./"]` is
provided as a default set of include paths, and is presumed as a value
if no other value is available. If the sanitized set of paths is empty,
SCSS is simply not run.

There was a related bug here with file extensions: SCSSLint handles
`.css` files in addition to `.scss`, but `#sanitize_include_paths` was
only looking for the `.scss` file extension. I've fixed that filtering
code to pull in the list of allowed extensions directly from SCSSLint.